fixed smart joystick slider rescale

+++ src/main/java/poroslib/triggers/SmartJoystick.java
@@ -13,10 +13,10 @@ public class SmartJoystick extends Joystick
 	private int twistAxis = 2;
 	
 	private int sliderAxis = 3;
-	private int oldMinSlider = 1;
-	private int oldMaxSlider = -1;
-	private int minSlider = 0;
-	private int maxSlider = 1;
+	private double oldMinSlider = 1;
+	private double oldMaxSlider = -1;
+	private double minSlider = 0;
+	private double maxSlider = 1;
 
 	public SmartJoystick(int port) 
 	{
@@ -80,7 +80,7 @@ public class SmartJoystick extends Joystick
 		this.twistAxis = twistAxis;
 	}
 	
-	public void SetSlider(int sliderAxis, int newMin, int newMax, int oldMin, int oldMax)
+	public void SetSlider(int sliderAxis, double newMin, double newMax, double oldMin, double oldMax)
 	{
 		this.sliderAxis = sliderAxis;
 		this.oldMinSlider = oldMin;
